refactor: split helpers out of main.py into focused modules
Extract models, validation, config, display, gh_client, and sync from main.py,
leaving main.py as the CLI/bootstrap/wiring entry point. Helper modules do not
import main. Re-export moved symbols from main for backwards compatibility and
propagate TOKEN/USE_COLORS/log patches via a custom module class.

- Add models.py for TypedDicts/dataclasses
- Add validation.py for sanitization and validators
- Add config.py for defaults, config loading, and runtime constants
- Add display.py for UI helpers, logging formatters, and AlertSystem
- Add gh_client.py for blocklist fetching and in-memory cache
- Add sync.py for folder/rule orchestration
- Trim main.py to CLI, argument parsing, and module wiring
- Update tests to patch canonical owning modules
- Update CI cov/mypy paths and docs (AGENTS.md, WARP.md, README.md)
