+++ AGENTS.md
@@ -4,9 +4,10 @@
 
 ### Project overview
 
-Single-file Python CLI tool (`main.py`) that syncs Control D DNS folders with
-remote JSON block-lists via the Control D REST API. No frontend, no database, no
-Docker required.
+Python CLI tool that syncs Control D DNS folders with remote JSON block-lists via
+the Control D REST API. The codebase is split into focused modules; `main.py` is
+only the CLI/bootstrap/wiring entry point. No frontend, no database, no Docker
+required.
 
 ### Runtime
 
@@ -22,7 +23,7 @@ Docker required.
 | -------------------------- | ------------------------------------------------------------------ |
 | Install/update deps        | `uv sync --all-extras`                                             |
 | Run linter                 | `ruff check .` (install via `uv tool install ruff` if not on PATH) |
-| Run all tests              | `uv run pytest tests/ -v`                                          |
+| Run all tests              | `uv run pytest tests/ test_main.py -v`                               |
 | Run specific test file     | `uv run pytest tests/test_fix_env.py -v`                           |
 | Run app (dry-run)          | `uv run python main.py --dry-run`                                  |
 | Run app (live)             | `TOKEN=xxx PROFILE=yyy uv run python main.py`                      |
@@ -31,9 +32,12 @@ Docker required.
 
 ### Current status
 
-`main.py` compiles cleanly when checked with the project interpreter
-(`uv run python -m py_compile main.py`). All 30 `test_*.py` modules under
-`tests/` are importable. Run the full test suite with `uv run pytest tests/ -v`.
+All modules compile cleanly (`uv run python -m py_compile main.py models.py
+validation.py config.py display.py gh_client.py sync.py api_client.py cache.py
+fix_env.py`). The full test suite is `uv run pytest tests/ test_main.py -v`.
+
+`main.py` is now only CLI/bootstrap/wiring; helpers live in `models.py`,
+`validation.py`, `config.py`, `display.py`, `gh_client.py`, and `sync.py`.
 
 ### Secrets for live runs
